requirements.txt
google-cloud-aiplatform==1.17.1
kfp==1.8.18
tfx==1.12.0
tensorflow==2.11.0
tensorflow-datasets==4.8.2
tensorflow-recommenders==0.7.2
tensor-annotations-tensorflow-stubs==2.0.2
kaggle==1.5.12
pandas==1.5.2
pandas-stubs==1.5.2.230105
tqdm==4.61.1
tqdm-stubs==0.2.1
